Authorization implemented

+++ src/api/notes/handler.js
@@ -15,14 +15,19 @@ class NotesHandler {
     async postNoteHandler(request, h) {
         try {
             this._validator.validateNotePayload(request.payload);
+
             const {
                 title = 'untitled', body, tags
             } = request.payload;
+            const {
+                id: credentialId
+            } = request.auth.credentials;
 
             const noteId = await this._service.addNote({
                 title,
                 body,
-                tags
+                tags,
+                owner: credentialId,
             });
 
             const response = h.response({
@@ -55,8 +60,13 @@ class NotesHandler {
         }
     }
 
-    async getNotesHandler() {
-        const notes = await this._service.getNotes();
+    async getNotesHandler(request) {
+        const {
+            id: credentialId
+        } = request.auth.credentials;
+
+        const notes = await this._service.getNotes(credentialId);
+
         return {
             status: 'success',
             data: {
@@ -70,7 +80,14 @@ class NotesHandler {
             const {
                 id
             } = request.params;
+            const {
+                id: credentialId
+            } = request.auth.credentials;
+
+            await this._service.verifyNoteOwner(id, credentialId);
+
             const note = await this._service.getNoteById(id);
+
             return {
                 status: 'success',
                 data: {
@@ -101,20 +118,16 @@ class NotesHandler {
     async putNoteByIdHandler(request, h) {
         try {
             this._validator.validateNotePayload(request.payload);
-            const {
-                title,
-                body,
-                tags
-            } = request.payload;
+
             const {
                 id
             } = request.params;
+            const {
+                id: credentialId
+            } = request.auth.credentials;
 
-            await this._service.editNoteById(id, {
-                title,
-                body,
-                tags
-            });
+            await this._service.verifyNoteOwner(id, credentialId);
+            await this._service.editNoteById(id, request.payload);
 
             return {
                 status: 'success',
@@ -146,6 +159,11 @@ class NotesHandler {
             const {
                 id
             } = request.params;
+            const {
+                id: credentialId
+            } = request.auth.credentials;
+
+            await this._service.verifyNoteOwner(id, credentialId);
             await this._service.deleteNoteById(id);
 
             return {
